Fix multiplication hash

+++ tables.hpp
@@ -36,7 +36,8 @@ unsigned long hash_index(bool division, unsigned m, double A, int key, int i){
 	if(division){
 		return (key+i) % m;
 	}else{
-		return floor((double)m * ((double)(key+i) * A) - floor((double)(key+i) * A));
+		return floor(  m * (  (float) (key+i)*A - floor((float) (key+i)*A)  )  );
+		//return floor((double)m * ((double)(key+i) * A) - floor((double)(key+i) * A));
 	}
 }
